+++ CHANGELOG.md
@@ -1,16 +1,19 @@
 # CHANGELOG
 
 ## [Unreleased]
-* List all expenses by id;
-* Update an expense;
 * Logic delete an expense;
 * Permanent delete an expense;
-* Create test services;
 * Add exception handler;
 * Configure docker;
 * Add Configure swagger;
 * Add tags;
 * Tag expenses;
+
+## [v1.3.0]() - 2020-12-26
+* Update an expense;
+* Test update expense on service layer;
+* Test update expense on controller layer;
+
 
 ## [v1.2.0]() - 2020-12-26
 ## Added